refactor: better movieClient types naming
