+++ jobs/Frontend/src/hooks/movies/useGetMovieDetails.ts
@@ -1,6 +1,9 @@
 import { useQuery } from '@tanstack/react-query'
 import movieClient from '../../api/movieClient/movieClient'
-import { FullMovieResponse, ApiError } from '../../api/movieClient/types'
+import {
+  FullMovieResponse,
+  ApiError,
+} from '../../api/movieClient/movieClientTypes'
 
 export const useGetMovieDetails = (movieId: string) =>
   useQuery<FullMovieResponse | ApiError, ApiError, FullMovieResponse>({
