+++ jobs/Frontend/src/hooks/movies/useGetMovies.ts
@@ -1,6 +1,10 @@
 import { useQuery } from '@tanstack/react-query'
 import movieClient from '../../api/movieClient/movieClient'
-import { ApiError, ApiResponse, Movie } from '../../api/movieClient/types'
+import {
+  ApiError,
+  ApiResponse,
+  Movie,
+} from '../../api/movieClient/movieClientTypes'
 
 export const useGetMovies = (query: string, page: number = 1) =>
   useQuery<ApiResponse<Movie> | ApiError, ApiError, ApiResponse<Movie>>({
